transform/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
confluent-kafka==1.7.0
ftfy==6.0.3
hvac==0.11.0
idna==3.2
mailchecker==4.0.11
phonenumbers==8.12.31
python-benedict==0.24.2
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
wcwidth==0.2.5
xmltodict==0.12.0